+++ sync-for-commerce/src/codatsynccommerce/sdk.py
@@ -6,28 +6,27 @@ from .connections import Connections
 from .integrations import Integrations
 from .sdkconfiguration import SDKConfiguration
 from .sync import Sync
-from .sync_flow_preferences import SyncFlowPreferences
+from .sync_flow_settings import SyncFlowSettings
 from codatsynccommerce import utils
 from codatsynccommerce.models import shared
 
 class CodatSyncCommerce:
     r"""Sync for Commerce: The API for Sync for Commerce.
 
-    Sync for Commerce is an API and a set of supporting tools built to enable e-commerce and point of sale platforms to provide high-quality integrations with numerous accounting platform through standardized API, seamlessly transforming business sale's data into accounting artefacts.
+    Sync for Commerce automatically replicates and reconciles sales data from a merchant’s source PoS, Payments, and eCommerce systems into their accounting software. This eliminates manual processing by merchants and transforms their ability to run and grow their business.
 
     [Read More...](https://docs.codat.io/commerce/overview)
 
-    <!-- Not seeing the end points you're expecting? We've reorganized our products, and you may be using a [different version of Sync for Commerce](https://docs.codat.io/sync-for-commerce-v1-api#/). -->
+    Not seeing the endpoints you're expecting? We've [reorganized our products](https://docs.codat.io/updates/230901-new-products), and you may be using a [different version of Sync for Commerce](https://docs.codat.io/sync-for-commerce-v1-api#/).
     """
     advanced_controls: AdvancedControls
     r"""Advanced company management and sync preferences."""
     connections: Connections
-    r"""Create new and manage existing Sync for Commerce companies using the Sync flow UI."""
     integrations: Integrations
     r"""View useful information about codat's integrations."""
     sync: Sync
     r"""Initiate and monitor the sync of company data into accounting software."""
-    sync_flow_preferences: SyncFlowPreferences
+    sync_flow_settings: SyncFlowSettings
     r"""Configure preferences for any given Sync for Commerce company using sync flow."""
 
     sdk_configuration: SDKConfiguration
@@ -70,5 +69,5 @@ class CodatSyncCommerce:
         self.connections = Connections(self.sdk_configuration)
         self.integrations = Integrations(self.sdk_configuration)
         self.sync = Sync(self.sdk_configuration)
-        self.sync_flow_preferences = SyncFlowPreferences(self.sdk_configuration)
+        self.sync_flow_settings = SyncFlowSettings(self.sdk_configuration)